Fix display of array items in Schema component
This commit changes the display of array items so that the actual sub-schema of
said items is displayed.

This reduces the asthetic annoyance when an empty "Array[]" schema is displayed
(when no sub-schema is given), and also fixes a bug where the description of
array items was not rendered at all.